Move Dto to domain and entity mappers to mapper class
