+++ app/src/main/java/com/example/briefly/data/remote/dto/NewsResultsDto.kt
@@ -1,8 +1,5 @@
 package com.example.briefly.data.remote.dto
 
-import com.example.briefly.domain.model.NewsByIdResponse
-import com.example.briefly.domain.model.NewsResponse
-
 data class NewsResponseDto(
     val response: NewsResultsDto
 )
@@ -22,20 +19,3 @@ data class NewsByIdContentDto(
     val status: String?,
     val total: Int?
 )
-
-fun NewsResponseDto.toNewsResponse() = with(response) {
-    NewsResponse(
-        news = results?.map { it.toNewsItem() } ?: emptyList(),
-        status = status ?: "",
-        totalResults = total ?: 0
-    )
-}
-
-fun NewsByIdResponseDto.toNewsByIdResponse() = with(response) {
-    NewsByIdResponse(
-        content = content?.toNewsItem(),
-        status = status ?: "",
-        totalResults = total ?: 0
-
-    )
-}
